Moved LeaderEochCheckpoint and LeaderEpochCheckPointFile to org.apache.kafka.storage.internals.checkpoint package.
