+++ checkstyle/import-control.xml
@@ -370,12 +370,16 @@
       <allow pkg="org.apache.kafka.server.log" />
       <allow pkg="org.apache.kafka.server.record" />
       <allow pkg="org.apache.kafka.test" />
-
+      <allow pkg="org.apache.kafka.storage"/>
       <subpackage name="remote">
         <allow pkg="scala.collection" />
       </subpackage>
 
     </subpackage>
+  </subpackage>
+
+  <subpackage name="storage.internals">
+    <allow pkg="org.apache.kafka.server"/>
   </subpackage>
 
   <subpackage name="shell">
