+++ core/src/test/scala/unit/kafka/log/LogTestUtils.scala
@@ -26,7 +26,7 @@ import kafka.utils.TestUtils
 import org.apache.kafka.common.Uuid
 import org.apache.kafka.common.record.{CompressionType, ControlRecordType, EndTransactionMarker, FileRecords, MemoryRecords, RecordBatch, SimpleRecord}
 import org.apache.kafka.common.utils.{Time, Utils}
-import org.apache.kafka.server.log.internals.{AbortedTxn, AppendOrigin,  FetchDataInfo, FetchIsolation, LazyIndex, LeaderEpochCheckpointFile, LogConfig, LogDirFailureChannel, TransactionIndex}
+import org.apache.kafka.server.log.internals.{AbortedTxn, AppendOrigin,  FetchDataInfo, FetchIsolation, LazyIndex, LogConfig, LogDirFailureChannel, TransactionIndex}
 import org.junit.jupiter.api.Assertions.{assertEquals, assertFalse}
 
 import java.nio.file.Files
@@ -34,6 +34,7 @@ import java.util.concurrent.{ConcurrentHashMap, ConcurrentMap}
 import kafka.log
 import org.apache.kafka.common.config.TopicConfig
 import org.apache.kafka.server.util.Scheduler
+import org.apache.kafka.storage.internals.checkpoint.LeaderEpochCheckpointFile
 
 import scala.jdk.CollectionConverters._
