+++ core/src/test/scala/unit/kafka/log/remote/RemoteLogManagerTest.scala
@@ -27,6 +27,7 @@ import org.apache.kafka.common.{KafkaException, TopicIdPartition, TopicPartition
 import org.apache.kafka.server.log.internals._
 import org.apache.kafka.server.log.remote.storage.RemoteStorageManager.IndexType
 import org.apache.kafka.server.log.remote.storage._
+import org.apache.kafka.storage.internals.checkpoint.LeaderEpochCheckpoint
 import org.apache.kafka.test.TestUtils
 import org.junit.jupiter.api.Assertions._
 import org.junit.jupiter.api.{BeforeEach, Test}
