+++ core/src/test/scala/unit/kafka/server/epoch/LeaderEpochFileCacheTest.scala
@@ -20,7 +20,8 @@ package kafka.server.epoch
 import kafka.utils.TestUtils
 import org.apache.kafka.common.TopicPartition
 import org.apache.kafka.common.requests.OffsetsForLeaderEpochResponse.{UNDEFINED_EPOCH, UNDEFINED_EPOCH_OFFSET}
-import org.apache.kafka.server.log.internals.{EpochEntry, LeaderEpochCheckpoint, LeaderEpochCheckpointFile, LeaderEpochFileCache}
+import org.apache.kafka.server.log.internals.{EpochEntry, LeaderEpochFileCache}
+import org.apache.kafka.storage.internals.checkpoint.{LeaderEpochCheckpoint, LeaderEpochCheckpointFile}
 import org.junit.jupiter.api.Assertions._
 import org.junit.jupiter.api.Test
